+++ src/components/Callout/Callout.stories.tsx
@@ -7,5 +7,6 @@ Template.displayName = Callout.displayName;
 
 export const Default: FictoanStory<typeof Callout> = createStoryFromTemplate(Template);
 Default.args = {
+    kind: "info",
     children: "Hello!"
 };